+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/TgFileUniqueId.kt
@@ -1,0 +1,15 @@
+package dev.inmo.tgbotapi.types
+
+import kotlinx.serialization.Serializable
+import kotlin.jvm.JvmInline
+
+@Serializable
+@JvmInline
+value class TgFileUniqueId(
+    val string: String
+)
+@Deprecated(
+    "Renamed",
+    ReplaceWith("TgFileUniqueId", "dev.inmo.tgbotapi.types.TgFileUniqueId")
+)
+typealias FileUniqueId = TgFileUniqueId