+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/AnimationFile.kt
@@ -2,7 +2,6 @@ package dev.inmo.tgbotapi.types.files
 
 import dev.inmo.tgbotapi.requests.abstracts.FileId
 import dev.inmo.tgbotapi.types.*
-import dev.inmo.tgbotapi.types.files.*
 import dev.inmo.tgbotapi.utils.MimeType
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -12,7 +11,7 @@ data class AnimationFile(
     @SerialName(fileIdField)
     override val fileId: FileId,
     @SerialName(fileUniqueIdField)
-    override val fileUniqueId: FileUniqueId,
+    override val fileUniqueId: TgFileUniqueId,
     override val width: Int,
     override val height: Int,
     override val duration: Long? = null,