+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/PathedFile.kt
@@ -10,7 +10,7 @@ data class PathedFile(
     @SerialName(fileIdField)
     override val fileId: FileId,
     @SerialName(fileUniqueIdField)
-    override val fileUniqueId: FileUniqueId,
+    override val fileUniqueId: TgFileUniqueId,
     @SerialName(filePathField)
     val filePath: String,
     @SerialName(fileSizeField)