+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/TelegramMediaFile.kt
@@ -2,7 +2,7 @@ package dev.inmo.tgbotapi.types.files
 
 import dev.inmo.tgbotapi.utils.internal.ClassCastsIncluded
 import dev.inmo.tgbotapi.requests.abstracts.FileId
-import dev.inmo.tgbotapi.types.FileUniqueId
+import dev.inmo.tgbotapi.types.TgFileUniqueId
 
 /**
  * Declare common part of media files in Telegram. Note: it is not representation of JVM `File` type
@@ -10,6 +10,6 @@ import dev.inmo.tgbotapi.types.FileUniqueId
 @ClassCastsIncluded
 sealed interface TelegramMediaFile {
     val fileId: FileId
-    val fileUniqueId: FileUniqueId
+    val fileUniqueId: TgFileUniqueId
     val fileSize: Long?
 }